+++ src/media/data/media/contents/movie/internal.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 #
-# Copyright 2023 Chris Josephes
+# Copyright 2024 Chris Josephes
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -29,9 +29,7 @@
 
 from datetime import timedelta
 from media.xml.namespaces import Namespaces
-from media.data.media.contents.generic.catalog import (
-        Title, TitleValueException, Catalog
-        )
+from media.data.media.contents import AbstractContent, ContentException
 from media.data.media.contents.generic.story import Story
 from media.data.media.contents.genericv.crew import Crew
 from media.data.media.contents.genericv.technical import Technical
@@ -39,16 +37,13 @@ from media.data.media.contents.movie.classification import Classification
 from media.generic.sorting.lists import ContentIndex
 
 
-class Movie():
+class Movie(AbstractContent):
     '''Movie object'''
-    def __init__(self, in_chunk):
-        self.title = None
-        self.catalog = None
+    def __init__(self, in_element):
+        super().__init__()
         self.technical = None
         self.crew = None
-        self.sort_title = ""
-        self.unique_key = ""
-        self._process(in_chunk)
+        self._process(in_element)
 
     def build_index_object(self):
         """
@@ -57,15 +52,9 @@ class Movie():
         """
         return MovieIndexEntry(self)
 
-    def _process(self, in_chunk):
-        for child in in_chunk:
-            if child.tag == Namespaces.nsf('movie') + 'title':
-                try:
-                    self.title = Title(child.text)
-                except TitleValueException as tve:
-                    raise MovieException(tve.message) from tve
-            if child.tag == Namespaces.nsf('movie') + 'catalog':
-                self.catalog = Catalog(child)
+    def _process(self, in_element):
+        super()._process(in_element)
+        for child in in_element:
             if child.tag == Namespaces.nsf('movie') + 'classification':
                 self.classification = Classification(child)
             if child.tag == Namespaces.nsf('movie') + 'technical':
@@ -76,33 +65,7 @@ class Movie():
                 self.story = Story(child)
             if child.tag == Namespaces.nsf('movie') + 'crew':
                 self.crew = Crew(child)
-        if self.title is not None:
-            self._build_unique_key()
-
-    def _build_unique_key(self):
-        ukv = None
-        cpy = None
-        if self.catalog is not None:
-            if self.catalog.alt_titles is not None:
-                if self.catalog.alt_titles.variant_sort is True:
-                    self.sort_title = \
-                            self.catalog.alt_titles.variant_title.sort_title
-                    self.unique_key = self.sort_title
-                else:
-                    self.sort_title = self.title.sort_title
-                    self.unique_key = self.sort_title
-            if self.catalog.copyright is not None:
-                cpy = str(self.catalog.copyright.year)
-            else:
-                cpy = "0000"
-            if self.catalog.unique_index is not None:
-                ukv = str(self.catalog.unique_index.index)
-            else:
-                ukv = "1"
-            self.unique_key += "-" + cpy + "-" + ukv
-        else:
-            self.sort_title = self.title.sort_title
-            self.unique_key = self.sort_title + "-0000-1"
+        self._post_load_process()
 
     def __hash__(self):
         return hash(self.unique_key)
@@ -157,7 +120,7 @@ class MovieIndexEntry(ContentIndex):
         self.first_letter = self.sort_title[0]
 
 
-class MovieException(Exception):
+class MovieException(ContentException):
     '''Exception raised when there is an issue with a movie.'''
     def __init__(self, message):
         super().__init__(message)